Removed Rest API from TimeOff Repository wo tests
