+++ .github/config.yml
@@ -12,6 +12,3 @@ requestInfoDefaultTitles:
 
 # *OPTIONAL* Label to be added to Issues and Pull Requests with insufficient information given
 requestInfoLabelToAdd: needs-more-info
-
-todo:
-  keyword: "TODO"
