+++ app/src/main/java/it/gym/repository/ReservationRepository.java
@@ -4,12 +4,13 @@ import it.gym.model.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.rest.webmvc.RepositoryRestController;
+import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Date;
 import java.util.List;
 
-@RepositoryRestController
+@Repository
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 
     List<Reservation> findByStartTime(Date startTime);
