+++ app/src/main/java/it/gym/repository/TimeOffRepository.java
@@ -3,13 +3,14 @@ import it.gym.model.TimeOff;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Date;
 import java.util.List;
 
 
-@RepositoryRestResource(path="/timesOff")
+@Repository
 public interface TimeOffRepository extends JpaRepository<TimeOff, Long> {
 
     @Query("select t from TimeOff as t where t.startTime >= :starttime  and t.endTime <= :endtime")
