+++ app/src/test/java/it/gym/service/GymServiceTest.java
@@ -1,6 +1,8 @@
 package it.gym.service;
 
 import it.gym.exception.GymNotFoundException;
+import it.gym.exception.InvalidReservationException;
+import it.gym.exception.InvalidTimeException;
 import it.gym.exception.NotFoundException;
 import it.gym.model.AUser;
 import it.gym.model.Customer;
@@ -17,8 +19,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.time.DayOfWeek;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.Locale;
 import java.util.Optional;
 
+import static org.apache.commons.lang3.time.DateUtils.addHours;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 
@@ -66,17 +72,79 @@ public class GymServiceTest {
         Mockito.verify(repository).delete(any(Gym.class));
     }
 
+    @Test
+    public void isValidInterval() {
+        Date start = new Date();
+        start = addHours(start, 1);
+        Date end = addHours(start, 1);
+
+        service.isValidInterval(start, end);
+    }
+
+    @Test(expected = InvalidReservationException.class)
+    public void startAfterEndTime() {
+        Date start = new Date();
+        start = addHours(start, 1);
+        Date end = addHours(start, -1);
+
+        service.isValidInterval(start, end);
+    }
+
+
+    @Test(expected = InvalidReservationException.class)
+    public void isIntervalPast() {
+        Date start = new Date();
+        start = addHours(start, -1);
+        Date end = addHours(start, 1);
+
+        service.isValidInterval(start, end);
+    }
+
+
+    @Test
+    public void isWithinWorkingHours() {
+        Calendar cal = Calendar.getInstance(Locale.ITALIAN);
+        cal.set(2019, Calendar.JUNE, 8, 12, 0);
+        Date start = cal.getTime();
+        Date end = addHours(start, 1);
+        Mockito.doReturn(Optional.of(createGym())).when(repository).findById(1L);
+        service.isWithinWorkingHours(1L, start, end);
+    }
+
+    @Test(expected = InvalidTimeException.class)
+    public void isNotWithinWorkingHours() {
+        Calendar cal = Calendar.getInstance(Locale.ITALIAN);
+        cal.set(2019, Calendar.JUNE, 8, 13, 0);
+        Date start = cal.getTime();
+        Date end = addHours(start, 1);
+        Mockito.doReturn(Optional.of(createGym())).when(repository).findById(1L);
+        service.isWithinWorkingHours(1L, start, end);
+    }
+
     private Gym createGym() {
         Gym gym = new Gym();
         gym.setId(1L);
         gym.setWeekStartsOn(DayOfWeek.MONDAY);
         gym.setMondayOpen(true);
-        gym.setTuesdayOpen(false);
-        gym.setWednesdayOpen(false);
-        gym.setThursdayOpen(false);
-        gym.setFridayOpen(false);
-        gym.setSaturdayOpen(false);
+        gym.setMondayStartHour(8);
+        gym.setMondayEndHour(22);
+        gym.setTuesdayStartHour(8);
+        gym.setTuesdayEndHour(22);
+        gym.setWednesdayStartHour(8);
+        gym.setWednesdayEndHour(22);
+        gym.setThursdayStartHour(8);
+        gym.setThursdayEndHour(22);
+        gym.setFridayStartHour(8);
+        gym.setFridayEndHour(22);
+        gym.setSaturdayStartHour(8);
+        gym.setSaturdayEndHour(13);
+        gym.setTuesdayOpen(true);
+        gym.setWednesdayOpen(true);
+        gym.setThursdayOpen(true);
+        gym.setFridayOpen(true);
+        gym.setSaturdayOpen(true);
         gym.setSundayOpen(false);
         return gym;
     }
+
 }
